% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{create_mouse_btn}
\alias{create_mouse_btn}
\title{Create directional arrow buttons}
\usage{
create_mouse_btn(direction)
}
\arguments{
\item{direction}{one of c("up", "down", "left", "right")}
}
\value{
button
}
\description{
Wrapper method to make directional buttons
}
