% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{move_mouse}
\alias{move_mouse}
\title{Move the mouse and click}
\usage{
move_mouse(direction)
}
\arguments{
\item{direction}{one of c("left", "right", "up", "down")}
}
\description{
Move mouse relative to "clicked" position
}
